requirements.txt
faiss_cpu==1.11.0
fastapi==0.115.12
groq==0.26.0
numpy==2.2.6
pandas==2.3.0
pydantic==2.11.5
python-dotenv==1.1.0
Requests==2.32.3
sentence_transformers==4.1.0
uvicorn==0.22.0
